WIP: Improving Makefile and added missing configs
Still in the process of fixing the build system. Running into a lot of
weird errors with multiple definitions in object files once I "fixed" STM32 drivers (see
README changes in the firmware-rebuild directory). As well as memory
region errors which I suspect to be from misconfiguration of either the
MCU configs or the way I am building.