Changed configs to fix build errors
Previously mentioned changes to STMs code in order to remove the build
errors in the README are now undone. Issues stemmed from the configs not
having the PCD module enabled and including the HAL header file. Also
got rid of the seemingly useless gcc flag, `-specs=nano.specs` that I
took from the older code. No documentation exists for that flag so idk
what it does.